Separated out dependency management into separate plugin.
