+++ build.sbt
@@ -1,11 +1,5 @@
-import sbt.Keys.resolvers
 addCommandAlias("makeSite", "docs/makeSite")
 addCommandAlias("console", "datasource/console")
-
-// NB: Make sure to update the Spark version in pyrasterframes/python/setup.py
-ThisBuild / rfSparkVersion := "2.3.2"
-ThisBuild / rfGeoTrellisVersion := "2.2.0"
-ThisBuild / rfGeoMesaVersion := "2.2.1"
 
 lazy val root = project
   .in(file("."))
@@ -28,14 +22,13 @@ lazy val core = project
   .disablePlugins(SparkPackagePlugin)
   .settings(
     moduleName := "rasterframes",
-    resolvers += "Azavea Public Builds" at "https://dl.bintray.com/azavea/geotrellis",
     libraryDependencies ++= Seq(
-      "com.chuusai" %% "shapeless" % "2.3.2",
-      "org.locationtech.jts" % "jts-core" % "1.16.0",
-      "org.locationtech.geomesa" %% "geomesa-z3" % rfGeoMesaVersion.value,
-      "org.locationtech.geomesa" %% "geomesa-spark-jts" % rfGeoMesaVersion.value exclude("jgridshift", "jgridshift"),
-      "com.azavea.geotrellis" %% "geotrellis-contrib-vlm" % "2.11.0",
-      "com.azavea.geotrellis" %% "geotrellis-contrib-gdal" % "2.11.0",
+      shapeless,
+      `jts-core`,
+      geomesa("z3").value,
+      geomesa("spark-jts").value,
+      `geotrellis-contrib-vlm`,
+      `geotrellis-contrib-gdal`,
       spark("core").value % Provided,
       spark("mllib").value % Provided,
       spark("sql").value % Provided,
@@ -46,7 +39,7 @@ lazy val core = project
         ExclusionRule(organization = "org.scalastic"),
         ExclusionRule(organization = "org.scalatest")
       ),
-      scalaTest
+      scalatest
     ),
     buildInfoKeys ++= Seq[BuildInfoKey](
       name, version, scalaVersion, sbtVersion, rfGeoTrellisVersion, rfGeoMesaVersion, rfSparkVersion
